requirements.txt
decorator==4.4.1
Django==3.0.2
django-rest-auth==0.9.5
djangorestframework==3.11.0
django-crispy-forms
gunicorn
pytz==2019.3
sqlparse==0.3.0
